Retired bash version of up completion.
Looking back, worked for only a limited set of folders, and the
task of figuring out what folder mapped with what suffix wasn't even
performed in the completion. It was shorter, but overall the python
version should prove to be an improvement.

I also commonized profile-appendix and bash_aliases under the resolution
that any system-specific stuff should just be in .profile anyway.